requirements.txt
Django==1.9.4
django-tastypie==0.13.3
python-dateutil==2.5.1
python-mimeparse==1.5.1
six==1.10.0
